requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

#azure-functions==1.11.0
azure-functions==1.11.3b1
azure-storage-blob